kurento-test: add isRunningContainer() coordination with .lock file
Have a look at commit:
"elastestbrowsers: use a lock file to guard entrypoint execution"
in the elastest project:
https://github.com/elastest/elastest-user-emulator-service

it introduces a .lock file in the Docker image entrypoint, that can be
used from this side of the code to know whether the entrypoint script
has finished working or not.

We should not start tests before the image's entrypoint has finished
starting up all services (PulseAudio, Xvfb, VNC), so this is a easy way
to do so. If the .lock file exists, then we have to wait; when the .lock
file is not found, it means that the entrypoint has finished.

Also a good thing from this technique is that it is compatible with
older images that didn't have the .lock file.

Also, move calls to startRecordingIfNeeded() after the container has
been fully started (when entrypoint has finished running).