kurento-test: spawn commands with a TTY to avoid buffering issues
When external commands are executed without a TTY, their stdout and
stderr outputs get buffered, with a maximum size of 4K. FFmpeg is able
to fill that buffer very easily, and when it does, the process ends
abruptly.

Fix this by allocating a virtual TTY for the process, then handling the
outbut buffers manually in our code.

Also trim the command output so leading spaces don't get in the way.