kurento-test: TEST: use a longer video for playback tests
The "small" video is only 5 seconds, and sometimes the browser takes
more than 2 or 3 seconds just to start up playback, so the total
playback time tends to be less than half of the whole video clip.

To offset the time taken for startup, it's safer to just use a longer
video. "rgb640x360.mov" is 15 seconds long, and shows a nice pattern of
Red, Gren, Blue colors that can be user for color detection.